Fix Technical task number Add package diagram Package refactoring
