+++ docs/component.puml
@@ -1,0 +1,28 @@
+@startuml
+
+package configuration {
+    package json {
+        [JsonUtils]
+    }
+
+    package route {
+        [Routable]
+    }
+
+    package stub {
+        [MockConfiguration]
+    }
+}
+
+package controller {
+    [AdminMockController] as cont
+}
+
+package manager {
+    [MockEndpointManager]
+}
+
+cont -> JsonUtils
+cont -> MockConfiguration
+
+@enduml
