+++ src/main/scala/paulymorph/mock/configuration/stub/MockConfiguration.scala
@@ -1,6 +1,4 @@
-package paulymorph.mock.configuration
-
-import paulymorph.mock.configuration.stub.Stub
+package paulymorph.mock.configuration.stub
 
 sealed trait MockConfiguration {
   def port: Int
