+++ src/main/scala/paulymorph/mock/controller/AdminMockConfigurationController.scala
@@ -1,4 +1,4 @@
-package paulymorph.mock.manager
+package paulymorph.mock.controller
 
 import java.util.concurrent.atomic.AtomicReference
 
@@ -10,19 +10,21 @@ import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Route
 import akka.stream.Materializer
 import com.typesafe.scalalogging.Logger
-import paulymorph.mock.configuration.MockConfiguration
+import paulymorph.mock.configuration.stub.MockConfiguration
+import paulymorph.mock.manager.MockEndpointManager
 import paulymorph.utils.Directives
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
-case class AdminMockConfigurationManager(adminPort: Int, endpointManager: MockEndpointManager)
-                                        (implicit actorSystem: ActorSystem,
-                                         materializer: Materializer) {
+case class AdminMockConfigurationController(adminPort: Int, endpointManager: MockEndpointManager)
+                                           (implicit actorSystem: ActorSystem,
+                                            materializer: Materializer) {
+
   import akka.http.scaladsl.server.Directives._
   import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
-  import paulymorph.mock.configuration.JsonUtils._
+  import paulymorph.mock.configuration.json.JsonUtils._
 
-  private implicit val executionContext = actorSystem.dispatcher
+  private implicit val executionContext: ExecutionContext = actorSystem.dispatcher
 
   val adminRoute: Route = pathPrefix("mock") {
     (post & entity(as[MockConfiguration])) { stub =>
@@ -30,14 +32,14 @@ case class AdminMockConfigurationManager(adminPort: Int, endpointManager: MockEn
         complete(StatusCodes.Created, s"Created a mock on port ${stub.port}")
       }
     } ~
-    pathPrefix(IntNumber) { port =>
-      get {
-        complete(endpointManager.getMock(port))
-      } ~
-      delete {
-        complete(endpointManager.deleteMock(port))
+      pathPrefix(IntNumber) { port =>
+        get {
+          complete(endpointManager.getMock(port))
+        } ~
+          delete {
+            complete(endpointManager.deleteMock(port))
+          }
       }
-    }
   }
 
   val swaggerRoute: Route = path("swagger") {
@@ -51,7 +53,7 @@ case class AdminMockConfigurationManager(adminPort: Int, endpointManager: MockEn
       .map(binding => atomicBinding.set(Some(binding)))
       .map(_ => ())
 
-  private val logger = Logger[AdminMockConfigurationManager]
+  private val logger = Logger[AdminMockConfigurationController]
 
   private val logDirective = Directives.logRequestResponse(logger)
